[SDPA-2689] Split preview and authenticate functions into separate libs.
